+++ app/src/main/java/io/jim/tesserapp/ui/controllers/Rotatable.kt
@@ -1,8 +1,8 @@
 package io.jim.tesserapp.ui.controllers
 
 interface Rotatable {
-    var x: Float
-    var y: Float
-    var z: Float
-    var q: Float
+    var x: Double
+    var y: Double
+    var z: Double
+    var q: Double
 }
