+++ app/src/test/java/io/jim/tesserapp/util/util.kt
@@ -4,14 +4,14 @@ import io.jim.tesserapp.math.vector.VectorN
 import junit.framework.Assert.assertEquals
 
 @Suppress("unused")
-fun assertEquals(expected: VectorN, actual: VectorN, delta: Float) {
+fun assertEquals(expected: VectorN, actual: VectorN, delta: Double) {
     expected.forEachIndexed { index, float ->
         assertEquals(float, actual[index], delta)
     }
 }
 
 @Suppress("unused")
-fun assertEquals(message: String, expected: VectorN, actual: VectorN, delta: Float) {
+fun assertEquals(message: String, expected: VectorN, actual: VectorN, delta: Double) {
     expected.forEachIndexed { index, float ->
         assertEquals(message, float, actual[index], delta)
     }
